microservicio-analisis/requirements.txt
flask==3.0.0
flask-cors==4.0.0
cx_Oracle==8.3.0
python-dotenv==1.0.0
reportlab==4.0.7
matplotlib==3.8.2
pandas==2.1.4
numpy==1.26.2